blog_website/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.3
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
